move notifier_sse
